requirements.txt
aiosignal==1.3.1
astroid==2.13.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
countryinfo==0.1.2
dill==0.3.6
frozenlist==1.3.3
google-i18n-address==2.5.2
idna==3.4
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==2.6.2
pylint==2.15.10
python-environ==0.4.54
python-slugify==7.0.0
requests==2.28.1
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.2
